Move log2 impl to float module Move ln() impl to float module
