--- src/hazmat/log2.rs
@@ -1,106 +1,0 @@
-// Copied from the `libm` crate, which is licensed under the MIT license.
-
-/* origin: FreeBSD /usr/src/lib/msun/src/e_log2.c */
-/*
- * ====================================================
- * Copyright (C) 1993 by Sun Microsystems, Inc. All rights reserved.
- *
- * Developed at SunSoft, a Sun Microsystems, Inc. business.
- * Permission to use, copy, modify, and distribute this
- * software is freely granted, provided that this notice
- * is preserved.
- * ====================================================
- */
-/*
- * Return the base 2 logarithm of x.  See log.c for most comments.
- *
- * Reduce x to 2^k (1+f) and calculate r = log(1+f) - f + f*f/2
- * as in log.c, then combine and scale in extra precision:
- *    log2(x) = (f - f*f/2 + r)/log(2) + k
- */
-
-const IVLN2HI: f64 = 1.44269504072144627571e+00; /* 0x3ff71547, 0x65200000 */
-const IVLN2LO: f64 = 1.67517131648865118353e-10; /* 0x3de705fc, 0x2eefa200 */
-const LG1: f64 = 6.666666666666735130e-01; /* 3FE55555 55555593 */
-const LG2: f64 = 3.999999999940941908e-01; /* 3FD99999 9997FA04 */
-const LG3: f64 = 2.857142874366239149e-01; /* 3FD24924 94229359 */
-const LG4: f64 = 2.222219843214978396e-01; /* 3FCC71C5 1D8E78AF */
-const LG5: f64 = 1.818357216161805012e-01; /* 3FC74664 96CB03DE */
-const LG6: f64 = 1.531383769920937332e-01; /* 3FC39A09 D078C69F */
-const LG7: f64 = 1.479819860511658591e-01; /* 3FC2F112 DF3E5244 */
-
-/// The base 2 logarithm of `x` (f64).
-pub fn log2(mut x: f64) -> f64 {
-    let x1p54 = f64::from_bits(0x4350000000000000); // 0x1p54 === 2 ^ 54
-
-    let mut ui: u64 = x.to_bits();
-    let hfsq: f64;
-    let f: f64;
-    let s: f64;
-    let z: f64;
-    let r: f64;
-    let mut w: f64;
-    let t1: f64;
-    let t2: f64;
-    let y: f64;
-    let mut hi: f64;
-    let lo: f64;
-    let mut val_hi: f64;
-    let mut val_lo: f64;
-    let mut hx: u32;
-    let mut k: i32;
-
-    hx = (ui >> 32) as u32;
-    k = 0;
-    if hx < 0x00100000 || (hx >> 31) > 0 {
-        if ui << 1 == 0 {
-            return -1. / (x * x); /* log(+-0)=-inf */
-        }
-        if (hx >> 31) > 0 {
-            return (x - x) / 0.0; /* log(-#) = NaN */
-        }
-        /* subnormal number, scale x up */
-        k -= 54;
-        x *= x1p54;
-        ui = x.to_bits();
-        hx = (ui >> 32) as u32;
-    } else if hx >= 0x7ff00000 {
-        return x;
-    } else if hx == 0x3ff00000 && ui << 32 == 0 {
-        return 0.;
-    }
-
-    /* reduce x into [sqrt(2)/2, sqrt(2)] */
-    hx += 0x3ff00000 - 0x3fe6a09e;
-    k += (hx >> 20) as i32 - 0x3ff;
-    hx = (hx & 0x000fffff) + 0x3fe6a09e;
-    ui = ((hx as u64) << 32) | (ui & 0xffffffff);
-    x = f64::from_bits(ui);
-
-    f = x - 1.0;
-    hfsq = 0.5 * f * f;
-    s = f / (2.0 + f);
-    z = s * s;
-    w = z * z;
-    t1 = w * (LG2 + w * (LG4 + w * LG6));
-    t2 = z * (LG1 + w * (LG3 + w * (LG5 + w * LG7)));
-    r = t2 + t1;
-
-    /* hi+lo = f - hfsq + s*(hfsq+R) ~ log(1+f) */
-    hi = f - hfsq;
-    ui = hi.to_bits();
-    ui &= (-1i64 as u64) << 32;
-    hi = f64::from_bits(ui);
-    lo = f - hi - hfsq + s * (hfsq + r);
-
-    val_hi = hi * IVLN2HI;
-    val_lo = (lo + hi) * IVLN2LO + lo * IVLN2HI;
-
-    /* spadd(val_hi, val_lo, y), except for not using double_t: */
-    y = k.into();
-    w = y + val_hi;
-    val_lo += (y - w) + val_hi;
-    val_hi = w;
-
-    val_lo + val_hi
-}
